requirements.txt
# LLM & Orchestration
langchain==0.1.20
langchain-google-genai==1.0.1
langgraph==0.0.26
langchain-core==0.1.23
pydantic==2.6.0

# Vector DB & Embeddings
chromadb==1.3.5
sentence-transformers==2.3.1

# Document Processing
arxiv==2.1.0
pypdf==4.0.1
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.1
numpy==1.24.3
pandas==2.0.3

# Search & Retrieval
rank-bm25==0.2.2

# UI
streamlit==1.31.0

# Testing
pytest==8.0.0